Support ECR image caching (#39)
## Description
This change updates our terraform module to optionally support the ECR caching that we've added on the agent side. Once we've done more testing, we can switch it to be on by default.

This change also makes the `gcp_project_id` parameter configurable to match the configuration for the cloudformation template.

## Testing
Testing completed successfully via applying in our sandbox account.