Rename "data plane" to "agent storage" in terraform (#46)
## Description
This change updates the agent terraform module to rename "data plane" to "agent storage" everywhere. It also switches the JWT public key to the prod value and makes it an input variable so we can override it for testing purposes.

See https://github.com/airplanedev/airport/pull/4110 for the associated changes on the airport side.

## Testing
Testing to be done by applying after merging the airport change.